requirements.txt
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.9.0
mediapipe>=0.10.9
numpy>=1.26.0
Pillow>=10.2.0
scipy>=1.12.0
scikit-image>=0.22.0
matplotlib>=3.8.0
tqdm>=4.66.0
wandb>=0.16.0
pyyaml>=6.0.1
einops>=0.7.0
lpips>=0.1.4
clean-fid>=0.1.35
albumentations>=1.3.1
